+++ NEWS.md
@@ -2,6 +2,12 @@
 
 ## Issues
 
+- Simplified input format for `location` arg in `ph_with_*` functions (#623). Instead of a `location` object 
+  created by the `ph_location_*` function family, `ph_with_*` functions now resolves certain short form
+  input into corresponding `location` objects. For example, instead of using `ph_location_label("<label>")`
+  you can now simply pass the `"<label>"` to the `location` arg. The functionn will convert the string into
+  the corresponding location object automatically. Other examples are `"body [1]"` for 
+  `ph_location_type(type = "body", type_idx = 1)`, or the integer `1` for `ph_location_id(id = 1)`
 - Add `slide_visible()` to get and set the visibility of slides (#622).
 - debug selector for `ph_remove()` (see #625) that was not working 
 for rvg outputs.